# backend_gateway/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PrepSense Gateway API",
...
    "version": "1.4.0",
    "contact": {
      "name": "PrepSense Development Team",
      "email": "dev@prepsense.com",
      "url": "https://github.com/dankimjw/PrepSense"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8001",
      "description": "Development server"
    },
    {
      "url": "http://localhost:8002",
      "description": "Testing server"
    },
    {
      "url": "https://api.prepsense.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/v1/users/": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "List Users",
        "description": "List all users (admin only)",
        "operationId": "list_users_api_v1_users__get",
        "security": [
          {
            "OAuth2PasswordBearerOptional": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserResponse"
                  },
                  "title": "Response List Users Api V1 Users  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{user_id}": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Read User",
        "description": "Get a specific user by ID",
        "operationId": "read_user_api_v1_users__user_id__get",
        "security": [
          {
            "OAuth2PasswordBearerOptional": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/recipes/recipe/{recipe_id}": {
      "get": {
        "tags": [
          "recipes",
          "Spoonacular Recipes"
        ],
        "summary": "Get recipe information",
        "description": "Get detailed information about a specific recipe.\nThis endpoint uses Spoonacular images when available, with local backups.",
        "operationId": "get_recipe_information_api_v1_recipes_recipe__recipe_id__get",
        "parameters": [
          {
            "name": "recipe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Recipe Id"
            }
          },
          {
            "name": "include_nutrition",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include nutrition information",
              "default": true,
              "title": "Include Nutrition"
            },
            "description": "Include nutrition information"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Recipe Information Api V1 Recipes Recipe  Recipe Id  Get"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/user-recipes": {
      "post": {
        "tags": [
          "recipes",
          "User Recipes"
        ],
        "summary": "Save a recipe to user's collection",
        "description": "Save a recipe to the user's collection",
        "operationId": "save_user_recipe_api_v1_user_recipes_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveRecipeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Save User Recipe Api V1 User Recipes Post"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "recipes",
          "User Recipes"
        ],
        "summary": "Get user's saved recipes (My Recipes)",
        "description": "Get user's saved recipes with optional filters. By default, excludes external recipes (Spoonacular) to keep My Recipes clean.",
        "operationId": "get_user_recipes_api_v1_user_recipes_get",
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by source",
              "title": "Source"
            },
            "description": "Filter by source"
          },
          {
            "name": "is_favorite",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by favorite status",
              "title": "Is Favorite"
            },
            "description": "Filter by favorite status"
          },
          {
            "name": "rating",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by rating",
              "title": "Rating"
            },
            "description": "Filter by rating"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by status: 'saved' or 'cooked'",
              "title": "Status"
            },
            "description": "Filter by status: 'saved' or 'cooked'"
          },
          {
            "name": "demo_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Filter to show only demo recipes",
              "default": false,
              "title": "Demo Only"
            },
            "description": "Filter to show only demo recipes"
          },
          {
            "name": "include_external",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include external recipes (Spoonacular, etc.) - defaults to False for My Recipes",
              "default": false,
              "title": "Include External"
            },
            "description": "Include external recipes (Spoonacular, etc.) - defaults to False for My Recipes"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "Maximum number of recipes to return",
              "default": 100,
              "title": "Limit"
            },
            "description": "Maximum number of recipes to return"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of recipes to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of recipes to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Get User Recipes Api V1 User Recipes Get"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/user-recipes/bookmarked-external": {
      "get": {
        "tags": [
          "recipes",
          "User Recipes"
        ],
        "summary": "Get user's bookmarked external recipes",
        "description": "Get user's bookmarked external recipes (Spoonacular, etc.) - separate from My Recipes",
        "operationId": "get_bookmarked_external_recipes_api_v1_user_recipes_bookmarked_external_get",
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "External source filter - defaults to 'spoonacular'",
              "default": "spoonacular",
              "title": "Source"
            },
            "description": "External source filter - defaults to 'spoonacular'"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "Maximum number of recipes to return",
              "default": 100,
              "title": "Limit"
            },
            "description": "Maximum number of recipes to return"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of recipes to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of recipes to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Get Bookmarked External Recipes Api V1 User Recipes Bookmarked External Get"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/user-recipes/stats": {
      "get": {
        "tags": [
          "recipes",
          "User Recipes"
        ],
        "summary": "Get user's recipe statistics",
        "description": "Get statistics about user's saved recipes (excludes external recipes from main counts)",
        "operationId": "get_recipe_stats_api_v1_user_recipes_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Recipe Stats Api V1 User Recipes Stats Get"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/api/v1/user-recipes/{recipe_id}/rating": {
      "put": {
        "tags": [
          "recipes",
          "User Recipes"
        ],
        "summary": "Update recipe rating",
        "description": "Update the rating for a saved recipe",
        "operationId": "update_recipe_rating_api_v1_user_recipes__recipe_id__rating_put",
        "parameters": [
          {
            "name": "recipe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Recipe Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRatingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Recipe Rating Api V1 User Recipes  Recipe Id  Rating Put"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/user-recipes/{recipe_id}/favorite": {
      "put": {
        "tags": [
          "recipes",
          "User Recipes"
        ],
        "summary": "Toggle recipe favorite status",
        "description": "Toggle the favorite status for a saved recipe",
        "operationId": "toggle_recipe_favorite_api_v1_user_recipes__recipe_id__favorite_put",
        "parameters": [
          {
            "name": "recipe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Recipe Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Toggle Recipe Favorite Api V1 User Recipes  Recipe Id  Favorite Put"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/user-recipes/{recipe_id}/mark-cooked": {
      "put": {
        "tags": [
          "recipes",
          "User Recipes"
        ],
        "summary": "Mark recipe as cooked",
        "description": "Mark a saved recipe as cooked (auto-promotion from saved to cooked status)",
        "operationId": "mark_recipe_as_cooked_api_v1_user_recipes__recipe_id__mark_cooked_put",
        "parameters": [
          {
            "name": "recipe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Recipe Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Mark Recipe As Cooked Api V1 User Recipes  Recipe Id  Mark Cooked Put"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/user-recipes/{recipe_id}": {
      "delete": {
        "tags": [
          "recipes",
          "User Recipes"
        ],
        "summary": "Delete a saved recipe",
        "description": "Delete a recipe from user's collection",
        "operationId": "delete_user_recipe_api_v1_user_recipes__recipe_id__delete",
        "parameters": [
          {
            "name": "recipe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Recipe Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete User Recipe Api V1 User Recipes  Recipe Id  Delete"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "recipes",
          "User Recipes"
        ],
        "summary": "Get a specific saved recipe",
        "description": "Get a specific recipe from user's collection",
        "operationId": "get_user_recipe_api_v1_user_recipes__recipe_id__get",
        "parameters": [
          {
            "name": "recipe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Recipe Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get User Recipe Api V1 User Recipes  Recipe Id  Get"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/demo/recipes": {
      "get": {
        "tags": [
          "admin",
          "Demo"
        ],
        "summary": "Get demo recipes for testing",
        "description": "Returns demo recipes formatted for the chat interface.\nThese recipes are designed to test various unit conversion scenarios.",
        "operationId": "get_demo_recipes_api_v1_demo_recipes_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Demo Recipes Api V1 Demo Recipes Get"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UpdateRatingRequest": {
        "properties": {
          "rating": {
            "type": "string",
            "title": "Rating",
            "description": "New rating: 'thumbs_up', 'thumbs_down', 'neutral'"
          }
        },
        "type": "object",
        "required": [
          "rating"
        ],
        "title": "UpdateRatingRequest",
        "description": "Request model for updating recipe rating"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token authentication"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key authentication for external services"
      }
    },
...
  },
...
}